pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradingagents"
version = "0.2.1"
description = "TradingAgents: Multi-Agents LLM Financial Trading Framework"
readme = "docs/README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Tauric Research"},
]
keywords = ["trading", "llm", "agents", "finance", "alpaca"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "alpaca-py>=0.40.0",
    "fastapi>=0.111.0",
    "langchain-core>=0.3.81",
    "langchain-anthropic>=0.3.15",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "numpy>=1.26.0",
    "pandas>=2.3.0",
    "parsel>=1.10.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "pytz>=2025.2",
    "questionary>=2.1.0",
    "rank-bm25>=0.2.2",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "typer>=0.21.0",
    "setuptools>=80.9.0",
    "stockstats>=0.6.5",
    "tqdm>=4.67.1",
    "typing-extensions>=4.14.0",
    "uvicorn>=0.30.0",
    "websockets>=12.0",
    "yfinance>=0.2.63",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "jupyter>=1.0.0",
]
dashboard = [
    "aiofiles>=23.2.0",
    "fastapi>=0.111.0",
    "pydantic>=2.7.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
]

[project.scripts]
tradingagents = "cli.main:app"

[project.urls]
Homepage = "https://github.com/TauricResearch/TradingAgents"
Documentation = "https://github.com/TauricResearch/TradingAgents/blob/main/docs/README.md"
Repository = "https://github.com/TauricResearch/TradingAgents"
Issues = "https://github.com/TauricResearch/TradingAgents/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["tradingagents*", "cli*"]

[tool.setuptools.package-data]
tradingagents = ["py.typed"]

# =============================================================================
# Linting & Formatting (Ruff)
# =============================================================================

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function calls in argument defaults
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# =============================================================================
# Type Checking (MyPy)
# =============================================================================

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

# =============================================================================
# Testing (Pytest)
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# =============================================================================
# Coverage
# =============================================================================

[tool.coverage.run]
source = ["src/tradingagents", "scripts"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
